restructure types folder
