+++ src/features/auth/api.ts
@@ -1,5 +1,5 @@
 import { apiClient } from "@/lib/axios/client";
-import { LoginRequest, LoginResponse } from "@/types";
+import { LoginRequest, LoginResponse } from "./types";
 
 export const authApi = {
   login: async (credentials: LoginRequest): Promise<LoginResponse> => {
