+++ src/features/auth/mutations.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 
-import { LoginRequest } from "@/types";
+import { LoginRequest } from "./types";
 import { authApi } from "./api";
 
 export const useLogin = () => {
