+++ src/features/categories/api.ts
@@ -1,10 +1,6 @@
 import { apiClient } from "@/lib/axios/client";
-import {
-  Category,
-  CreateCategoryRequest,
-  PaginatedResponse,
-  PaginationParams,
-} from "@/types";
+import { PaginatedResponse, PaginationParams } from "@/types";
+import { Category, CreateCategoryRequest } from "./types";
 
 export const categoriesApi = {
   findAll: async (
