+++ src/types/paginations.ts
@@ -1,0 +1,11 @@
+export interface PaginationParams {
+  page?: number;
+  limit?: number;
+}
+
+export interface PaginatedResponse<T> {
+  items: T[];
+  totalCount: number;
+  currentPage: number;
+  itemsPerPage: number;
+}
